Cleanup & Moq data
