+++ Nuix.Api/Nuix.Api/AutomapperProfiles/InvestmentAutoMapperProfile.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
 using Nuix.Data.Dto;
 using Nuix.Data.Model;
