+++ Nuix.Api/Nuix.Data/Dto/DetailInvestmentDto.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Nuix.Data.Dto
+﻿namespace Nuix.Data.Dto
 {
     public class DetailInvestmentDto
     {
