+++ Nuix.Api/Nuix.Data/Model/Investment.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Nuix.Data.Model
 {
