+++ Nuix.Api/Nuix.Data/Model/User.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Nuix.Data.Model
+﻿namespace Nuix.Data.Model
 {
     public class User
     {
