+++ Nuix.Api/Nuix.Data/Repository/IRepository.cs
@@ -1,9 +1,7 @@
-﻿using System;
+﻿using Nuix.Data.Model;
+using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using Nuix.Data.Model;
 
 namespace Nuix.Data.Repository
 {
