+++ Nuix.Api/Nuix.Data/Repository/MoqData/IData.cs
@@ -1,0 +1,10 @@
+﻿using Nuix.Data.Model;
+using System.Collections.Generic;
+
+namespace Nuix.Data.Repository.MoqData
+{
+    public interface IData
+    {
+        Dictionary<string, IEnumerable<Investment>> ClientFactory();
+    }
+}
